Semi-Intergrated Two Dimensional Slugs with LTLMoP
Cost can be enter into a Cost section in the .spec file
Formalas are entered in the format: d.d formula
Where d.d is the cost value for the formula (Ie the cost if the formula
is true) and formula is a GR(1) LTL formula that can contain the
following
All GR(1) Operators
Any region names
Any sensor names
Any actuator names

A small example specification has been added to demo this. See
/examples/optimalNegotiation/oneRobot/. The sitation is a robot who
wants starbucks but doesn't want to go outside if its snowing and who
refuses to wait in lines for coffee.